binaries: rename 'tt binaries' to 'tt binaries list'
Renamed 'tt binaries' to 'tt binaries list'.
Moved list module to binary module.

Part of #628
