+++ cli/binary/testdata/bin/tarantool
