+++ cli/binary/testdata/tarantool_dev/bin/tarantool
