+++ cli/binary/testdata/tarantool_dev/bin_symlink_broken/tarantool
