+++ cli/cmd/binaries.go
@@ -2,31 +2,36 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/tarantool/tt/cli/binary"
 	"github.com/tarantool/tt/cli/cmdcontext"
-	"github.com/tarantool/tt/cli/list"
 	"github.com/tarantool/tt/cli/modules"
 )
 
 // NewBinariesCmd creates binaries command.
 func NewBinariesCmd() *cobra.Command {
 	var binariesCmd = &cobra.Command{
-		Use:   "binaries",
+		Use: "binaries",
+	}
+
+	var listCmd = &cobra.Command{
+		Use:   "list",
 		Short: "Show a list of installed binaries and their versions.",
 		Run: func(cmd *cobra.Command, args []string) {
 			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
-				internalBinariesModule, args)
+				internalListModule, args)
 			handleCmdErr(cmd, err)
 		},
 	}
 
+	binariesCmd.AddCommand(listCmd)
 	return binariesCmd
 }
 
-// internalBinariesModule is a default binaries module.
-func internalBinariesModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
+// internalListModule is a list module.
+func internalListModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	if !isConfigExist(cmdCtx) {
 		return errNoConfig
 	}
 
-	return list.ListBinaries(cmdCtx, cliOpts)
+	return binary.ListBinaries(cmdCtx, cliOpts)
 }
