+++ cli/cmd/instances.go
@@ -2,8 +2,8 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
+	"github.com/tarantool/tt/cli/binary"
 	"github.com/tarantool/tt/cli/cmdcontext"
-	"github.com/tarantool/tt/cli/list"
 	"github.com/tarantool/tt/cli/modules"
 )
 
@@ -30,7 +30,7 @@ func internalInstancesModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 		return errNoConfig
 	}
 
-	err := list.ListInstances(cmdCtx, cliOpts)
+	err := binary.ListInstances(cmdCtx, cliOpts)
 
 	return err
 }
